Change add method to accept double parameters

+++ src/main/java/com/example/handson/Calculator.java
@@ -1,6 +1,6 @@
 package com.example.handson;
 public class Calculator {
-  public int add(int a, int b) {
+  public double add(double a, double b) {
     return a + b;
   }
 }
